add neuro

+++ main.py
@@ -2,7 +2,11 @@ from PyQt5.QtWidgets import QApplication, QMainWindow, QWidget, QFileDialog
 from PyQt5.QtGui import QPainter, QPen, QBrush, QColor, QPixmap
 from PyQt5.QtCore import Qt, QRectF, QSizeF
 from pyui.test_4 import Ui_MainWindow
+from pyui.neuro_1 import Ui_MainWindow_n
 from PIL import Image
+from Unet import *
+from Pspnet import *
+from funcs import *
 import sys
 import os
 
@@ -69,6 +73,7 @@ class App(QMainWindow, Ui_MainWindow):
         self.setupUi(self)
 
         self.have_photo = None
+        self.type_of_f = ''
 
         self.draw_objects = Drawing(QColor(0, 128, 0))
         self.draw_def = Drawing(QColor(0, 128, 0))
@@ -77,6 +82,7 @@ class App(QMainWindow, Ui_MainWindow):
         self.pushButton.clicked.connect(self.clear_objects)
         self.pushButton_2.clicked.connect(self.clear_defects)
         self.pushButton_30.clicked.connect(self.upload_photo)
+        self.pushButton_4.clicked.connect(self.get_from_neuro)
 
         self.colorButton_3.clicked.connect(lambda: self.draw_objects.change_color(QColor(0, 128, 0)))
         self.colorButton_2.clicked.connect(lambda: self.draw_objects.change_color(QColor(0, 0, 255)))
@@ -124,11 +130,11 @@ class App(QMainWindow, Ui_MainWindow):
                     QFileDialog.getSaveFileName(self, "Сохранить изображения", '',
                                                 'Изображение (*.png);;Изображение (*.jpg);')[
                         0]
-                type_of_f = filename.split('/')[-1].split('.')[-1]
-                self.draw_objects.save_as_image(f'objects_1.{type_of_f}')
-                self.draw_def.save_as_image(f'defects_1.{type_of_f}')
-                image_obj = Image.open(f'objects_1.{type_of_f}')
-                image_def = Image.open(f'defects_1.{type_of_f}')
+                self.type_of_f = filename.split('/')[-1].split('.')[-1]
+                self.draw_objects.save_as_image(f'objects_1.{self.type_of_f}')
+                self.draw_def.save_as_image(f'defects_1.{self.type_of_f}')
+                image_obj = Image.open(f'objects_1.{self.type_of_f}')
+                image_def = Image.open(f'defects_1.{self.type_of_f}')
                 image_start = Image.open(f'image_start.png')
                 image_object_size = image_obj.size
                 new_image = Image.new('RGB', (image_object_size[0], 3 * image_object_size[1] + 3), (255, 255, 255))
@@ -137,11 +143,61 @@ class App(QMainWindow, Ui_MainWindow):
                 new_image.paste(image_def, (0, 2 * image_object_size[1] + 2))
                 new_image.save(filename)
                 new_image.show()
-                os.remove(f'objects_1.{type_of_f}')
-                os.remove(f'defects_1.{type_of_f}')
-                os.remove('image_start.png')
             except Exception as e:
                 pass
+
+    def get_from_neuro(self):
+        if self.have_photo:
+            img = image.load_img('image_start.png', target_size=(img_height // 8, img_width // 8))
+            self.adm = Neuro(img)
+            self.adm.show()
+
+    def closeEvent(self, event):
+        if self.have_photo:
+            os.remove('image_start.png')
+        if self.type_of_f:
+            os.remove(f'objects_1.{self.type_of_f}')
+            os.remove(f'defects_1.{self.type_of_f}')
+
+
+class Neuro(QMainWindow, Ui_MainWindow_n):
+    def __init__(self, img):
+        super().__init__()
+        self.setupUi(self)
+        self.img = img
+        self.get_result()
+
+    def get_result(self):
+        modelUnet = unet(num_classes, (img_height // 8, img_width // 8, 3))
+        modelUnet.load_weights('Model_UNET_elem.h5')
+        modelPsPnet = pspnet(num_classes_def, (img_height // 8, img_width // 8, 3))
+        modelPsPnet.load_weights('Model_PSPNET_def.h5')
+
+        x = image.img_to_array(self.img)
+
+        predict_obj = np.array(modelUnet.predict(x.reshape(1, img_height // 8, img_width // 8, 3)))[0].reshape(-1,
+                                                                                                               num_classes)
+        predict_def = np.array(modelPsPnet.predict(x.reshape(1, img_height // 8, img_width // 8, 3)))[0].reshape(-1,
+                                                                                                                 num_classes_def)
+        pr1 = []
+        pr2 = []
+        for k in range(len(predict_obj)):
+            pr1.append(index2color(predict_obj[k]))
+            pr2.append(index2color_def(predict_def[k]))
+        pr1 = np.array(pr1)
+        pr2 = np.array(pr2)
+        pr1 = pr1.reshape(img_height // 8, img_width // 8, 3)
+        pr2 = pr2.reshape(img_height // 8, img_width // 8, 3)
+        res_obj = Image.fromarray(pr1.astype('uint8')).resize((1024, 64))
+        res_obj.save('image_neuro_obj.png')
+        res_def = Image.fromarray(pr2.astype('uint8')).resize((1024, 64))
+        res_def.save('image_neuro_def.png')
+        self.canvas_7.setPixmap(QPixmap('image_neuro_obj.png'))
+        self.canvas_8.setPixmap(QPixmap('image_neuro_def.png'))
+
+    def closeEvent(self, event):
+        os.remove('image_neuro_obj.png')
+        os.remove('image_neuro_def.png')
 
 
 def except_hook(cls, exception, traceback):
